:art: refactor(#182/auth): CookieInstruction을 api 모듈로 이동
