+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/auth/dto/result/AuthResult.java
@@ -1,9 +1,8 @@
 package org.sopt36.ninedotserver.auth.dto.result;
 
 import java.util.Optional;
-import org.sopt36.ninedotserver.auth.support.CookieInstruction;
 
 public sealed interface AuthResult permits LoginResult, SignupResult {
 
-    Optional<CookieInstruction> refreshTokenCookie();
+    Optional<String> refreshToken();
 }
