+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/auth/dto/result/SignupResult.java
@@ -2,7 +2,6 @@ package org.sopt36.ninedotserver.auth.dto.result;
 
 import java.util.Optional;
 import org.sopt36.ninedotserver.auth.model.ProviderType;
-import org.sopt36.ninedotserver.auth.support.CookieInstruction;
 
 public record SignupResult(
     ProviderType provider,
@@ -10,7 +9,7 @@ public record SignupResult(
     String name,
     String email,
     String picture,
-    Optional<CookieInstruction> refreshTokenCookie
+    Optional<String> refreshToken
 ) implements AuthResult {
 
 }
