+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/auth/service/login/dto/IssuedTokens.java
@@ -1,10 +1,8 @@
 package org.sopt36.ninedotserver.auth.service.login.dto;
 
-import org.sopt36.ninedotserver.auth.support.CookieInstruction;
-
 public record IssuedTokens(
     String accessToken,
-    CookieInstruction refreshTokenCookie
+    String refreshToken
 ) {
 
 }
